+++ flexiapi/resources/views/account/recovery/show.blade.php
@@ -39,14 +39,18 @@
                         </p>
                     @endif
                     <div>
-                        <input placeholder="+123456789" name="phone" type="text" value="{{ old('phone') }}">
+                        <input placeholder="+123456789" name="phone" type="text" value="@if ($phone){{ $phone }}@else{{ old('phone') }}@endif">
                         <label for="phone">{{ __('Phone number') }}</label>
                         @include('parts.errors', ['name' => 'phone'])
                         @include('parts.errors', ['name' => 'identifier'])
                     </div>
                 @endif
 
-                @include('parts.captcha')
+                @if (!empty($account_recovery_token))
+                    <input name="account_recovery_token" type="hidden" value="{{ $account_recovery_token }}">
+                @else
+                    @include('parts.captcha')
+                @endif
 
                 <div class="large">
                     <input class="btn oppose" type="submit" value="{{ __('Send')}}">