+++ flexiapi/resources/views/mails/recover_by_code.blade.php
@@ -9,7 +9,7 @@ You are trying to authenticate to {{ $account->space->name }} using your email a
 
 Please enter the code bellow to finish the authentication process.
 
-## {{ $account->recovery_code }
+## {{ $account->recovery_code }}
 
 @if (config('app.recovery_code_expiration_minutes') > 0)
 The code is only available {{ config('app.recovery_code_expiration_minutes') }} minutes.